+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ and by implication, [Semantic Versioning](http://semver.org/).
 - Updated get_checklist to return the checklist owner.
 - Updated get_checklist to return a list of other participants.
 - Updated get_checklist to return the number of observers.
+- Updated get_checklist to use table for extracting protocol values.
 
 [0.2.0] - 2024-10-23
 --------------------